requirements_dev.txt
absl-py==1.3.0
astunparse==1.6.3
cachetools==5.2.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
crypto==1.4.1
cryptography==38.0.3
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Deprecated==1.2.13
easydict==1.10
firebase==3.0.1
Flask==2.1.0
Flask-Reuploaded==0.3.2
fonttools==4.25.0
gast==0.3.3
gcloud==0.18.3
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
h5py==2.10.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
jws==0.1.3
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1653274189334/work
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib @ file:///C:/b/abs_ae02atcfur/croot/matplotlib-suite_1667356722968/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
munkres==1.1.4
Naked==0.1.31
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.1.1.26
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
Pillow==9.2.0
ply==3.11
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
PyQt5==5.15.7
PyQt5-sip @ file:///C:/Windows/Temp/abs_d7gmd2jg8i/croots/recipe/pyqt-split_1659273064801/work/pyqt_sip
Pyrebase4==4.5.0
pytesseract==0.3.10
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-jwt==2.0.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.4.1
shellescape==3.8.1
sip @ file:///C:/Windows/Temp/abs_b8fxd17m2u/croots/recipe/sip_1659012372737/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sseclient==0.0.27
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==2.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///C:/ci/tornado_1662476933490/work
tqdm==4.64.1
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3==1.26.12
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.10.0